requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.2.1
fastapi==0.115.12
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.5
pydantic_core==2.33.2
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
thread==2.0.5
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.3.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
